Move state.py to core
QtileState is only used by Qtile and Qtile is in core. I don't see a
reason why a normal user would want to tinker with QtileState.
