+++ libqtile/core/state.py
